mobile(P1.1b): fix ambiguous 'money' call in sum() — real build error
Human's first xcodebuild attempt against Money.swift failed: 'Ambiguous
use of money'. sum()'s empty-list branch called money(0, currency) with
a bare integer literal 0, which conforms to ExpressibleByIntegerLiteral
for BOTH the money(Int64,_) and money(Double,_) throws overloads --
Swift can't pick one. Fixed by writing Int64(0) explicitly. Every other
money() call site in the file already passed an already-typed
expression (roundHalfAwayFromZero's Int64 return, .amount fields, etc.)
so this was the only ambiguous spot -- confirmed by re-grepping the
whole file before committing.

+++ apps/ios/Domain/Sources/Domain/Money.swift
@@ -154,7 +154,12 @@ public func sum(_ items: [Money], currency: String? = nil) throws -> Money {
         guard let currency else {
             throw MoneyError(description: "sum() of empty list needs a currency")
         }
-        return money(0, currency)
+        // Int64(0), not the bare literal 0 -- an untyped integer literal is
+        // ambiguous between the money(Int64,_) and money(Double,_) throws
+        // overloads (both conform to ExpressibleByIntegerLiteral), which
+        // Swift reports as "Ambiguous use of 'money'" -- a real error the
+        // first xcodebuild attempt against this file surfaced.
+        return money(Int64(0), currency)
     }
     var acc = items[0]
     for m in items.dropFirst() {
